extract bucket and resource variable provider implementation to a base class
